requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
click==6.7
dominate==2.3.1
Flask==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.2
get==1.0.2
gunicorn==19.8.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
post==1.0.1
public==1.0.1
query-string==1.0.1
request==1.0.1
visitor==0.1.3
Werkzeug==0.14.1
wtf==0.1
WTForms==2.1
